+++ resources/lang/de/people.php
@@ -116,4 +116,5 @@ return [
     'user_removed_coupons_from_project' => '<strong>:user</strong> entfernte <strong>:amount :coupons</strong> von <strong>:project</strong> für <strong>:date</strong> hinzu',
     'import_people_data' => 'Personendaten importieren',
     'import_bank_data' => 'Bankdaten importieren',
+    'search_family_members' => 'Familienmitglieder suchen',
 ];