+++ resources/views/bank/person-card.blade.php
@@ -1,6 +1,10 @@
 <div class="card mb-4 bg-light">
     @php
         $frequentVisitor = $person->frequentVisitor;
+        $family = $person->family;
+        if (count($family) > 1) {
+            $family_filter = implode(' OR ', collect($family)->pluck('search')->sort()->toArray());
+        }
     @endphp
     <div class="card-header p-2" @if($frequentVisitor) style="background:lightgoldenrodyellow;" @endif >
         <div class="form-row">
@@ -113,4 +117,9 @@
             <em class="pb-2 px-2">@lang('people.no_coupons_defined')</em>
         @endforelse
     </div>
+    @if(isset($family_filter) && $filter != $family_filter)
+        <div class="card-footer p-2">
+            <a href="{{ route('bank.withdrawalSearch') }}?filter={{ $family_filter }}" class="btn btn-sm btn-success">@lang('people.search_family_members')</a>
+        </div>
+    @endif
 </div>